Module google_adexchangeseller2::api

Structs

Account
There is no detailed description.

AccountAdclientListCall
List all ad clients in this Ad Exchange account.

AccountAlertListCall
List the alerts for this Ad Exchange account.

AccountCustomchannelGetCall
Get the specified custom channel from the specified ad client.

AccountCustomchannelListCall
List all custom channels in the specified ad client for this Ad Exchange account.

AccountGetCall
Get information about the selected Ad Exchange account.

AccountListCall
List all accounts available to this Ad Exchange account.

AccountMetadataDimensionListCall
List the metadata for the dimensions available to this AdExchange account.

AccountMetadataMetricListCall
List the metadata for the metrics available to this AdExchange account.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeSeller hub.

AccountPreferreddealGetCall
Get information about the selected Ad Exchange Preferred Deal.

AccountPreferreddealListCall
List the preferred deals for this Ad Exchange account.

AccountReportGenerateCall
Generate an Ad Exchange report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

AccountReportSavedGenerateCall
Generate an Ad Exchange report based on the saved report ID sent in the query parameters.

AccountReportSavedListCall
List all saved reports in this Ad Exchange account.

AccountUrlchannelListCall
List all URL channels in the specified ad client for this Ad Exchange account.

Accounts
There is no detailed description.

AdClient
There is no detailed description.

AdClients
There is no detailed description.

AdExchangeSeller
Central instance to access all AdExchangeSeller related resource activities

Alert
There is no detailed description.

Alerts
There is no detailed description.

CustomChannel
There is no detailed description.

CustomChannelTargetingInfo
The targeting information of this custom channel, if activated.

CustomChannels
There is no detailed description.

Metadata
There is no detailed description.

PreferredDeal
There is no detailed description.

PreferredDeals
There is no detailed description.

Report
There is no detailed description.

ReportHeaders
The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

ReportingMetadataEntry
There is no detailed description.

SavedReport
There is no detailed description.

SavedReports
There is no detailed description.

UrlChannel
There is no detailed description.

UrlChannels
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.