Struct google_adexchangeseller2::AccountMethods

pub struct AccountMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeSeller hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangeseller2 as adexchangeseller2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangeseller2::AdExchangeSeller;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeSeller::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `adclients_list(...)`, `alerts_list(...)`, `customchannels_get(...)`, `customchannels_list(...)`, `get(...)`, `list(...)`, `metadata_dimensions_list(...)`, `metadata_metrics_list(...)`, `preferreddeals_get(...)`, `preferreddeals_list(...)`, `reports_generate(...)`, `reports_saved_generate(...)`, `reports_saved_list(...)` and `urlchannels_list(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

fn reports_saved_generate(&self, account_id: &str, saved_report_id: &str) -> AccountReportSavedGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an Ad Exchange report based on the saved report ID sent in the query parameters.

Arguments

accountId - Account owning the saved report.

savedReportId - The saved report to retrieve.

fn metadata_metrics_list(&self, account_id: &str) -> AccountMetadataMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

List the metadata for the metrics available to this AdExchange account.

Arguments

accountId - Account with visibility to the metrics.

fn reports_generate(&self, account_id: &str, start_date: &str, end_date: &str) -> AccountReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an Ad Exchange report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Arguments

accountId - Account which owns the generated report.

startDate - Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

endDate - End of the date range to report on in "YYYY-MM-DD" format, inclusive.

fn urlchannels_list(&self, account_id: &str, ad_client_id: &str) -> AccountUrlchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all URL channels in the specified ad client for this Ad Exchange account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to list URL channels.

fn reports_saved_list(&self, account_id: &str) -> AccountReportSavedListCall<'a, C, A>

Create a builder to help you perform the following task:

List all saved reports in this Ad Exchange account.

Arguments

accountId - Account owning the saved reports.

fn preferreddeals_get(&self, account_id: &str, deal_id: &str) -> AccountPreferreddealGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about the selected Ad Exchange Preferred Deal.

Arguments

accountId - Account owning the deal.

dealId - Preferred deal to get information about.

fn adclients_list(&self, account_id: &str) -> AccountAdclientListCall<'a, C, A>

Create a builder to help you perform the following task:

List all ad clients in this Ad Exchange account.

Arguments

accountId - Account to which the ad client belongs.

fn metadata_dimensions_list(&self, account_id: &str) -> AccountMetadataDimensionListCall<'a, C, A>

Create a builder to help you perform the following task:

List the metadata for the dimensions available to this AdExchange account.

Arguments

accountId - Account with visibility to the dimensions.

fn customchannels_get(&self, account_id: &str, ad_client_id: &str, custom_channel_id: &str) -> AccountCustomchannelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the specified custom channel from the specified ad client.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client which contains the custom channel.

customChannelId - Custom channel to retrieve.

fn customchannels_list(&self, account_id: &str, ad_client_id: &str) -> AccountCustomchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all custom channels in the specified ad client for this Ad Exchange account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to list custom channels.

fn list(&self) -> AccountListCall<'a, C, A>

Create a builder to help you perform the following task:

List all accounts available to this Ad Exchange account.

fn alerts_list(&self, account_id: &str) -> AccountAlertListCall<'a, C, A>

Create a builder to help you perform the following task:

List the alerts for this Ad Exchange account.

Arguments

accountId - Account owning the alerts.

fn preferreddeals_list(&self, account_id: &str) -> AccountPreferreddealListCall<'a, C, A>

Create a builder to help you perform the following task:

List the preferred deals for this Ad Exchange account.

Arguments

accountId - Account owning the deals.

fn get(&self, account_id: &str) -> AccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about the selected Ad Exchange account.

Arguments

accountId - Account to get information about. Tip: 'myaccount' is a valid ID.