Struct google_adexchangeseller2::AccountReportGenerateCall

pub struct AccountReportGenerateCall<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

Generate an Ad Exchange report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

This method supports media download. To enable it, adjust the builder like this: .param("alt", "media"). Please note that due to missing multi-part support on the server side, you will only receive the media, but not the Report structure that you would usually get. The latter will be a default value.

A builder for the reports.generate method supported by a account resource. It is not used directly, but through a AccountMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().reports_generate("accountId", "startDate", "endDate")
             .start_index(4)
             .add_sort("no")
             .add_metric("labore")
             .max_results(62)
             .locale("dolore")
             .add_filter("invidunt")
             .add_dimension("aliquyam")
             .doit();

Methods

impl<'a, C, A> AccountReportGenerateCall<'a, C, A> where
C: BorrowMut<Client>,
A: GetToken,

pub fn doit(self) -> Result<(Response, Report)>

Perform the operation you have build so far.

pub fn account_id(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

Account which owns the generated report.

Sets the account id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn start_date(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

Sets the start date query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn end_date(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

End of the date range to report on in "YYYY-MM-DD" format, inclusive.

Sets the end date query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn start_index(self, new_value: u32) -> AccountReportGenerateCall<'a, C, A>

Index of the first row of report data to return.

Sets the start index query property to the given value.

pub fn add_sort(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

The name of a dimension or metric to sort the resulting report on, optionally prefixed with "+" to sort ascending or "-" to sort descending. If no prefix is specified, the column is sorted ascending.

Append the given value to the sort query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn add_metric(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

Numeric columns to include in the report.

Append the given value to the metric query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn max_results(self, new_value: u32) -> AccountReportGenerateCall<'a, C, A>

The maximum number of rows of report data to return.

Sets the max results query property to the given value.

pub fn locale(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

Optional locale to use for translating report output to a local language. Defaults to "en_US" if not specified.

Sets the locale query property to the given value.

pub fn add_filter(self, new_value: &str) -> AccountReportGenerateCall<'a, C, A>

Filters to be run on the report.

Append the given value to the filter query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn add_dimension(
self,
new_value: &str
) -> AccountReportGenerateCall<'a, C, A>

Dimensions to base the report on.

Append the given value to the dimension query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> AccountReportGenerateCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> AccountReportGenerateCall<'a, C, A> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

pub fn add_scope<T, S>(self, scope: T) -> AccountReportGenerateCall<'a, C, A> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::AdexchangeSellerReadonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.