Struct google_adexchangeseller2::Report

pub struct Report {
    pub headers: Option<Vec<ReportHeaders>>,
    pub kind: Option<String>,
    pub rows: Option<Vec<Vec<String>>>,
    pub total_matched_rows: Option<String>,
    pub warnings: Option<Vec<String>>,
    pub averages: Option<Vec<String>>,
    pub totals: Option<Vec<String>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reports saved generate accounts (response)

reports generate accounts (response)

Fields

headers

The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

kind

Kind this is, in this case adexchangeseller#report.

rows

The output rows of the report. Each row is a list of cells; one for each dimension in the request, followed by one for each metric in the request. The dimension cells contain strings, and the metric cells contain numbers.

total_matched_rows

The total number of rows matched by the report request. Fewer rows may be returned in the response due to being limited by the row count requested or the report row limit.

warnings

Any warnings associated with generation of the report.

averages

The averages of the report. This is the same length as any other row in the report; cells corresponding to dimension columns are empty.

totals

The totals of the report. This is the same length as any other row in the report; cells corresponding to dimension columns are empty.