Struct google_adexchangeseller2::ReportingMetadataEntry

pub struct ReportingMetadataEntry {
    pub required_metrics: Option<Vec<String>>,
    pub kind: Option<String>,
    pub compatible_metrics: Option<Vec<String>>,
    pub compatible_dimensions: Option<Vec<String>>,
    pub id: Option<String>,
    pub required_dimensions: Option<Vec<String>>,
    pub supported_products: Option<Vec<String>>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
required_metrics: Option<Vec<String>>

The names of the metrics which the dimension or metric this reporting metadata entry describes requires to also be present in order for the report to be valid. Omitting these will not cause an error or warning, but may result in data which cannot be correctly interpreted.

kind: Option<String>

Kind of resource this is, in this case adexchangeseller#reportingMetadataEntry.

compatible_metrics: Option<Vec<String>>

The names of the metrics the dimension or metric this reporting metadata entry describes is compatible with.

compatible_dimensions: Option<Vec<String>>

For metrics this is a list of dimension IDs which the metric is compatible with, for dimensions it is a list of compatibility groups the dimension belongs to.

id: Option<String>

Unique identifier of this reporting metadata entry, corresponding to the name of the appropriate dimension or metric.

required_dimensions: Option<Vec<String>>

The names of the dimensions which the dimension or metric this reporting metadata entry describes requires to also be present in order for the report to be valid. Omitting these will not cause an error or warning, but may result in data which cannot be correctly interpreted.

supported_products: Option<Vec<String>>

The codes of the projects supported by the dimension or metric this reporting metadata entry describes.