Struct google_adexchangeseller2::SavedReports

pub struct SavedReports {
    pub next_page_token: Option<String>,
    pub items: Option<Vec<SavedReport>>,
    pub kind: Option<String>,
    pub etag: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reports saved list accounts (response)

Fields

next_page_token

Continuation token used to page through saved reports. To retrieve the next page of results, set the next request's "pageToken" value to this.

items

The saved reports returned in this list response.

kind

Kind of list this is, in this case adexchangeseller#savedReports.

etag

ETag of this response for caching purposes.