Struct google_adexchangeseller2::PreferredDeal

pub struct PreferredDeal {
    pub kind: Option<String>,
    pub buyer_network_name: Option<String>,
    pub currency_code: Option<String>,
    pub advertiser_name: Option<String>,
    pub fixed_cpm: Option<String>,
    pub start_time: Option<String>,
    pub end_time: Option<String>,
    pub id: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

preferreddeals get accounts (response)

Fields

kind

Kind of resource this is, in this case adexchangeseller#preferredDeal.

buyer_network_name

The name of the buyer network this deal is for.

currency_code

The currency code that applies to the fixed_cpm value. If not set then assumed to be USD.

advertiser_name

The name of the advertiser this deal is for.

fixed_cpm

The fixed price for this preferred deal. In cpm micros of currency according to currencyCode. If set, then this preferred deal is eligible for the fixed price tier of buying (highest priority, pay exactly the configured fixed price).

start_time

Time when this deal becomes active in seconds since the epoch (GMT). If not set then this deal is active immediately upon creation.

end_time

Time when this deal stops being active in seconds since the epoch (GMT). If not set then this deal is valid until manually disabled by the publisher.

id

Unique identifier of this preferred deal.